refactor(home): replace game card raster images with SVG vector assets (#1557)
